chore:refractor the constants extracted from the controllers
